More stubs with some findings...
We will want to pass in the drawerView to the interaction so that we can animate specific properties or views within the drawerView. Since the consumer is passing in the drawerView they should have control over what they animate in when opening and closing by passing an appropriate closure. Also we should only deal with open and close for states right now to avoid over complicating animations and handling too many states and the transitions between those states.